requirements.txt
asgiref==3.8.1
boto3==1.36.4
botocore==1.36.4
Django==5.1.5
django-cors-headers==4.6.0
django-rest-framework==0.1.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
jmespath==1.0.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.11.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
